// tui-library/projects/ng-tui/src/tui-core/interface/config.interface.ts
import { Item } from './item.interface';

export interface TUICoreConfig {
    iconfontPrefix: string;
    iconfontSymbolPrefix: string;
    iconfontLoadingClass: string[];
    buttonLoadingIcon: string;
    defaultColor: string;
    confirmCancelTitle: string;
    confirmOkTitle: string;
    errorImgSrc: string;
    weekTitles: string[];
    monthTitles: string[];
    pickerConfirmTitle: string;
    pickerCancleTitle: string;
    pickerNowTitle: string;
    pickerNowYearTitle: string;
    pickerCleanTitle: string;
    paginationItems: Item[];
    paginationLimitTextFormat: (limit: number) => string;
    uploadItemSize: number;
    tableConfig: TableConfig;
}

export interface TUIConfig {
    iconfontPrefix?: string;
    iconfontSymbolPrefix?: string;
    iconfontLoadingClass?: string[];
    buttonLoadingIcon?: string;
    defaultColor?: string;
    confirmCancelTitle?: string;
    confirmOkTitle?: string;
    errorImgSrc?: string;
    weekTitles?: string[];
    monthTitles?: string[];
    pickerConfirmTitle?: string;
    pickerCancleTitle?: string;
    pickerNowTitle?: string;
    pickerNowYearTitle?: string;
    pickerCleanTitle?: string;
    paginationItems?: Item[];

    paginationLimitTextFormat?: (limit: number) => string;

    uploadItemSize?: number;

    tableConfig?: TableConfig;
}

export interface TableConfig {
    emptyTitle: string;
    goTitle: string;
}